Extract cheffs dir behavior, move data bag classes to own files
