+++ lib/chef/chef_fs/file_system/repository/chef_repository_file_system_root_dir.rb
@@ -20,7 +20,7 @@ require "chef/chef_fs/file_system/base_fs_dir"
 require "chef/chef_fs/file_system/repository/chef_repository_file_system_acls_dir"
 require "chef/chef_fs/file_system/repository/chef_repository_file_system_cookbooks_dir"
 require "chef/chef_fs/file_system/repository/chef_repository_file_system_cookbook_artifacts_dir"
-require "chef/chef_fs/file_system/repository/chef_repository_file_system_data_bags_dir"
+require "chef/chef_fs/file_system/repository/data_bags_dir"
 require "chef/chef_fs/file_system/repository/chef_repository_file_system_entry"
 require "chef/chef_fs/file_system/repository/chef_repository_file_system_policies_dir"
 require "chef/chef_fs/file_system/repository/chef_repository_file_system_versioned_cookbooks_dir"
@@ -175,7 +175,7 @@ class Chef
             when "policies"
               dirs = paths.map { |path| ChefRepositoryFileSystemPoliciesDir.new(name, self, path) }
             when "data_bags"
-              dirs = paths.map { |path| ChefRepositoryFileSystemDataBagsDir.new(name, self, path) }
+              dirs = paths.map { |path| DataBagsDir.new(name, self, path) }
             when "acls"
               dirs = paths.map { |path| ChefRepositoryFileSystemAclsDir.new(name, self, path) }
             else
